create retriever file and move functions to where they are used
